Azure cosmos read operations (#1787)
* Cosmos DB read operations.

* Added some fixes for update/delete and partition key logic

* PartitionKey can be number, boolean or string. Updated @since

* Fixed issue with binding params for raw query

* Added missing test and refactor code to remove Sonar/compiler warnings

* Added tests for more custom result processing

* Added more tests with custom type results

+++ data-azure-cosmos/src/test/groovy/io/micronaut/data/azure/CosmosBasicSpec.groovy
@@ -14,10 +14,12 @@ import com.azure.cosmos.util.CosmosPagedIterable
 import com.fasterxml.jackson.databind.node.ObjectNode
 import io.micronaut.context.ApplicationContext
 import io.micronaut.core.type.Argument
+import io.micronaut.core.util.StringUtils
 import io.micronaut.data.azure.entities.Address
 import io.micronaut.data.azure.entities.Child
 import io.micronaut.data.azure.entities.CosmosBook
 import io.micronaut.data.azure.entities.Family
+import io.micronaut.data.azure.repositories.CosmosBookDtoRepository
 import io.micronaut.data.azure.repositories.CosmosBookRepository
 import io.micronaut.data.azure.repositories.FamilyRepository
 import io.micronaut.data.cosmos.config.CosmosDatabaseConfiguration
@@ -42,6 +44,8 @@ class CosmosBasicSpec extends Specification implements AzureCosmosTestProperties
     ApplicationContext context = ApplicationContext.run(properties)
 
     CosmosBookRepository bookRepository = context.getBean(CosmosBookRepository)
+
+    CosmosBookDtoRepository bookDtoRepository = context.getBean(CosmosBookDtoRepository)
 
     FamilyRepository familyRepository = context.getBean(FamilyRepository)
 
@@ -156,6 +160,41 @@ class CosmosBasicSpec extends Specification implements AzureCosmosTestProperties
             optFamily2.get().children.size() == 2
             optFamily2.get().address
         when:
+            def lastOrderedLastName = familyRepository.lastOrderedLastName()
+        then:
+            StringUtils.isNotEmpty(lastOrderedLastName)
+        when:
+            optFamily2.get().registeredDate = new Date()
+            familyRepository.update(optFamily2.get())
+            def lastOrderedRegisteredDate = familyRepository.lastOrderedRegisteredDate()
+        then:
+            lastOrderedRegisteredDate
+            lastOrderedRegisteredDate == optFamily2.get().registeredDate
+        when:
+            def exists = familyRepository.existsById(FAMILY1_ID)
+        then:
+            exists
+        when:
+            exists = familyRepository.existsById(UUID.randomUUID().toString())
+        then:
+            !exists
+        when:
+            exists = familyRepository.existsByIdAndRegistered(FAMILY1_ID, true)
+        then:
+            !exists
+        when:
+            exists = familyRepository.existsByIdAndRegistered(FAMILY1_ID, false)
+        then:
+            exists
+        when:
+            def cnt = familyRepository.count()
+        then:
+            cnt >= 2
+        when:
+            cnt = familyRepository.countByRegistered(false)
+        then:
+            cnt >= 2
+        when:"Using raw query for update is not supported"
             familyRepository.updateLastName(FAMILY1_ID, "New Last Name")
         then:
             thrown(IllegalStateException)
@@ -199,7 +238,7 @@ class CosmosBasicSpec extends Specification implements AzureCosmosTestProperties
         then:
             optFamily2.get().children.size() == 3
             optFamily1.get().address.state == "NY"
-        when:
+        when:"Using raw query for delete is not supported"
             familyRepository.deleteByRegistered(false)
         then:
             thrown(IllegalStateException)
@@ -248,6 +287,29 @@ class CosmosBasicSpec extends Specification implements AzureCosmosTestProperties
             optFamily1 = familyRepository.findById(FAMILY1_ID)
         then:
             !optFamily1.present
+    }
+
+    def "test DTO entity retrieval"() {
+        given:
+            CosmosBook book = new CosmosBook()
+            book.id = UUID.randomUUID().toString()
+            book.title = "New Book"
+            book.totalPages = 500
+            bookRepository.save(book)
+        when:
+            def loadedBook = bookRepository.queryById(book.id)
+        then:
+            loadedBook
+        when:
+            def bookDto = bookDtoRepository.findById(book.id)
+        then:
+            bookDto.present
+            bookDto.get().title == book.title
+            bookDto.get().totalPages == 500
+        when:
+            def bookDtos = bookDtoRepository.findByTitleAndTotalPages(book.title, book.totalPages)
+        then:
+            bookDtos.size() > 0
     }
 
     def "should get cosmos client"() {
